+++ R/predict_setting_contacts.R
@@ -88,7 +88,10 @@ predict_setting_contacts <- function(population,
 
   combination <- Reduce("+", setting_matrices)
   setting_matrices$all <- combination
-  setting_matrices$all <- new_age_matrix(setting_matrices$all)
+  setting_matrices$all <- new_age_matrix(
+    matrix = setting_matrices$all,
+    age_breaks = age_breaks
+  )
 
   # if we haven't set anything for the per capita household size, return this
   # adjusted matrix
@@ -104,7 +107,9 @@ predict_setting_contacts <- function(population,
     )
   }
 
-  setting_matrices <- new_setting_prediction_matrix(setting_matrices)
+  setting_matrices <- new_setting_prediction_matrix(setting_matrices,
+    age_breaks = age_breaks
+  )
 
   return(setting_matrices)
 }