+++ R/predictions_to_matrix.R
@@ -6,6 +6,7 @@
 #'
 #' @param contact_predictions data frame with columns `age_group_from`,
 #'   `age_group_to`, and `contacts`.
+#' @param ... extra arguments
 #'
 #' @return Square matrix with the unique age groups from `age_group_from/to`
 #'   in the rows and columns and `contacts` as the values.
@@ -27,7 +28,12 @@
 #' predictions_to_matrix(fairfield_school_contacts)
 #'
 #' @export
-predictions_to_matrix <- function(contact_predictions) {
+predictions_to_matrix <- function(contact_predictions, ...) {
+  UseMethod("predictions_to_matrix")
+}
+
+#' @export
+predictions_to_matrix.predicted_contacts <- function(contact_predictions, ...) {
   prediction_matrix <- contact_predictions %>%
     tidyr::pivot_wider(
       names_from = age_group_from,
@@ -37,7 +43,7 @@ predictions_to_matrix <- function(contact_predictions) {
       "age_group_to"
     ) %>%
     as.matrix() %>%
-    new_age_matrix()
+    new_age_matrix(age_breaks = age_breaks(contact_predictions))
 
   prediction_matrix
 }